+++ eglswapbuffers/eglswapbuffers.rs
@@ -70,7 +70,7 @@ struct Cli {
 
     ///Verbose.
     #[clap(short, long, action=clap::ArgAction::Count)]
-    verbose: usize,
+    verbose: u8,
 
     ///Show raw count
     #[clap(short = 'r', long)]
